requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
APScheduler==3.11.1
asyncpg==0.31.0
attrs==25.4.0
bitarray==3.8.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.1
cryptography==46.0.3
cytoolz==1.1.0
distro==1.9.0
docstring_parser==0.17.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.122.0
Flask==3.1.2
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
influxdb-client==1.49.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-core==1.1.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.49
lxml==6.0.2
MarkupSafe==3.0.3
mcp==1.22.0
multidict==6.7.0
ngrok==1.6.0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
parsimonious==0.10.0
pillow==12.0.0
propcache==0.4.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pyunormalize==17.0.0
PyYAML==6.0.3
reactivex==4.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rlp==4.1.0
rpds-py==0.29.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.50.0
tenacity==9.1.2
toolz==1.1.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
web3==7.14.0
websockets==15.0.1
Werkzeug==3.1.4
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
